+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - There can only be one instance of the `Publish Window` in the `Microservice Manager`
 
+### Changed
+- The `Publish Window` is now centered on show relative to the editor
+
 ### [1.6.0]
 ### Added
 - Displaying log pagination if message contains more that 5000 chars